已实现timeout功能，即可以指定每个方法的timeout时间，超过这个时间时，将返回相应的错误报文。

+++ rop/src/main/java/com/rop/annotation/ServiceMethod.java
@@ -56,7 +56,7 @@ public @interface ServiceMethod {
      * 为0或负数则表示不进行过期限制
      * @return
      */
-    long timeout() default -100;
+    int timeout() default -100;
 
     /**
      * 该方法所对应的版本号，对应version请求参数的值，版本为空，表示不进行版本限定
