+++ rop/src/main/java/com/rop/annotation/ServiceMethodGroup.java
@@ -46,7 +46,7 @@ public @interface ServiceMethodGroup {
      * 为0或负数则表示不进行过期限制
      * @return
      */
-    long timeout() default 0;
+    int timeout() default 0;
 
     /**
      * 该方法所对应的版本号，对应version请求参数的值，版本为空，表示不进行版本限定
